requirements.txt
fpdf
boto3
python-dotenv